requirements.txt
fastapi[all]
python-fasthtml
redis
python-jose[cryptography]
httpx-oauth
pytest
pytest-asyncio
python-dotenv
requests
httpx
pytest-mock
pydantic
pydantic-ai